@@ -2,16 +2,16 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 
 	<ItemGroup>
 
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Orleans.Client" Version="4.0.0-preview2" />
-		<PackageReference Include="Microsoft.Orleans.Sdk" Version="4.0.0-preview2" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.0-rc.2.22472.3" />
+		<PackageReference Include="Microsoft.Orleans.Client" Version="7.0.0-rc2" />
+		<PackageReference Include="Microsoft.Orleans.Sdk" Version="7.0.0-rc2" />
 	</ItemGroup>
 
 	<ItemGroup>
