@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Orleans.Sdk" Version="4.0.0-preview2" />
-		<PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="4.0.0-preview2" />
+		<PackageReference Include="Microsoft.Orleans.Sdk" Version="7.0.0-rc2" />
+		<PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="7.0.0-rc2" />
 	</ItemGroup>
 </Project>
