@@ -1,13 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="4.0.0-preview2" />
-		<PackageReference Include="Microsoft.Orleans.Sdk" Version="4.0.0-preview2" />
+		<PackageReference Include="Microsoft.Orleans.Runtime" Version="7.0.0-rc2" />
+		<PackageReference Include="Microsoft.Orleans.Sdk" Version="7.0.0-rc2" />
 	</ItemGroup>
 	<ItemGroup>
 	  <ProjectReference Include="..\OrleansStarterTemplate.Grains.Interfaces\OrleansStarterTemplate.Grains.Interfaces.csproj" />
