@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <PackageType>Template</PackageType>
-        <PackageVersion>0.0.1-alpha002</PackageVersion>
+        <PackageVersion>0.0.1-alpha003</PackageVersion>
         <PackageId>sparreio.Templates.OrleansStarterTemplate</PackageId>
         <PackageLicenseExpression>MIT</PackageLicenseExpression>
         <Title>Orleans Starter Template</Title>
