@@ -1,11 +1,10 @@
 ﻿using Orleans;
 
-namespace OrleansStarterTemplate.Grains.Interfaces
-{
-    public interface IHelloWorldGrain : IGrainWithStringKey
-    {
-        public const string Key = "HelloWorld";
+namespace OrleansStarterTemplate.Grains.Interfaces;
 
-        Task<HelloMessage> Hello(string name);
-    }
+public interface IHelloWorldGrain : IGrainWithStringKey
+{
+    public const string Key = "HelloWorld";
+
+    Task<HelloMessage> Hello(string name);
 }
